requirements.txt
alembic==1.16.5
annotated-types==0.7.0
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.6
ecdsa==0.19.1
greenlet==3.2.4
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.15.0